+++ zou/app/services/entities_service.py
@@ -1,9 +1,12 @@
-from zou.app.utils import events
 from zou.app.services import base_service
+from zou.app.utils import events
+from zou.app.utils import fields
 
 from zou.app.models.entity import Entity, EntityLink
 from zou.app.models.entity_type import EntityType
 from zou.app.models.preview_file import PreviewFile
+from zou.app.models.task import association_table as assignees_table
+from zou.app.models.task import Task
 
 from zou.app.services.exception import (
     PreviewFileNotFoundException,
@@ -108,3 +111,79 @@ def get_entity_links_for_project(project_id):
         .filter(Entity.project_id == project_id) \
         .all()
     return Entity.serialize_list(result)
+
+
+def get_entities_and_tasks(criterions={}):
+    """
+    Get all entities for given criterions with related tasks for each entity.
+    """
+    entity_map = {}
+    task_map = {}
+
+    query = (
+        Entity.query
+        .outerjoin(Task, Task.entity_id == Entity.id)
+        .outerjoin(assignees_table)
+        .add_columns(
+            Task.id,
+            Task.task_type_id,
+            Task.task_status_id,
+            Task.priority,
+            assignees_table.columns.person
+        )
+    )
+
+    if "entity_type_id" in criterions:
+        query = (
+            query.filter(Entity.entity_type_id == criterions["entity_type_id"])
+        )
+
+    if "project_id" in criterions:
+        query = query.filter(Entity.project_id == criterions["project_id"])
+
+    for (
+        entity,
+        task_id,
+        task_type_id,
+        task_status_id,
+        task_priority,
+        person_id
+    ) in query.all():
+        entity_id = str(entity.id)
+
+        entity.data = entity.data or {}
+
+        if entity_id not in entity_map:
+            entity_map[entity_id] = {
+                "id": str(entity.id),
+                "name": entity.name,
+                "description": entity.description,
+                "frame_in": entity.data.get("frame_in", None),
+                "frame_out": entity.data.get("frame_out", None),
+                "fps": entity.data.get("fps", None),
+                "preview_file_id": str(entity.preview_file_id or ""),
+                "canceled": entity.canceled,
+                "data": fields.serialize_value(entity.data),
+                "tasks": []
+            }
+
+        if task_id is not None:
+
+            if task_id not in task_map:
+                task_dict = {
+                    "id": str(task_id),
+                    "entity_id": entity_id,
+                    "task_status_id": str(task_status_id),
+                    "task_type_id": str(task_type_id),
+                    "priority": task_priority or 0,
+                    "assignees": []
+                }
+                task_map[task_id] = task_dict
+                entity_dict = entity_map[entity_id]
+                entity_dict["tasks"].append(task_dict)
+
+            if person_id:
+                task_map[task_id]["assignees"].append(str(person_id))
+
+    return list(entity_map.values())
+
